refactor: BaseEntity 패키지 변경
