+++ src/main/java/aegis/server/domain/common/domain/BaseEntity.java
@@ -1,5 +1,4 @@
-package aegis.server.domain.common;
-
+package aegis.server.domain.common.domain;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
