+++ src/main/java/aegis/server/domain/member/domain/Member.java
@@ -1,6 +1,6 @@
 package aegis.server.domain.member.domain;
 
-import aegis.server.domain.common.BaseEntity;
+import aegis.server.domain.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
